backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
